requirements.txt
pytest~=7.4.4
selene~=2.0.0rc4
allure-pytest